HMSymbol 优化资源解析模块，告警处理 Signed-off-by: dengyanfang23 <dengyanfang@huawei.com>

+++ rosen/modules/2d_graphics/include/text/hm_symbol_config_ohos.h
@@ -27,7 +27,7 @@ class HmSymbolConfigOhosImpl;
 
 class DRAWING_API HmSymbolConfigOhos {
 public:
-    static DrawingSymbolLayersGroups GetSymbolLayersGroups(uint32_t glyphId);
+    static DrawingSymbolLayersGroups GetSymbolLayersGroups(uint16_t glyphId);
 
     static std::vector<std::vector<DrawingPiecewiseParameter>> GetGroupParameters(
         DrawingAnimationType type, uint16_t groupSum, uint16_t animationMode = 0,
